updated streamlit_petmatch to show pictures and started on saving results to a file. currently it only appends the headers though..
